Extract LoopRenderer and update new way of block preview
